reoganize literate specs into smaller logical essays
